[Testing] Merge test changes from `permission_provider`.
