+++ tests/modules/ceb_test/ceb_test.info.yml
@@ -1,4 +1,4 @@
-name: ceb test Entity
+name: CEB Test Entity
 type: module
 description: Provides a bundle-able CEB test content entity used to store CEB test content.
 core: 8.x
